Remove libdrmdecrypt from AOSP build.
Not all boards that use the omap4 chipset need to have the drm
library.  In particular, the panda does not.  The libdrmdecrypt
make file was moved from hardware/ti/omap4xxx, to vendor/ti/proprietary.

The directory libdrmdecrypt was removed in this change, and added
in change Icbeb6c5c.

The dependency on libdrmdecrypt was removed from omap4.mk and added
to tuna_common.mk in change Ia1019321.

Change-Id: Iaad4d1122809ad67869d7515c5c5009a048bb1c2