OMAP4:Panda: Use old hwcomposer API version for panda
The hwcomposer for OMAP4 is generic to the all OMAP4
processors.

Panda does not support Vsync protocal which is expected
when the HAL declares an API support level of 0.3.

Therefore indicate to the upper layers that if the product is
panda then only API level 0.2 is supported which then forces
upper layers to fake the vsync signal

Change-Id: I22a39f55ea017ced25424556ffc95b099d0a2781
Signed-off-by: Dan Murphy <dmurphy@ti.com>