omx: Fix for message free from mmosal pipe
To append new message at the head of the pipe, the
pipe is read and written back. The temporary message
created during this step was freed unconditionly and
this was unnecessarily throwing error traces in the log.
This is fixed by freeing the temporary pipe memory only when
it was allocated.

Change-Id: Iaa432ddb9e4a612fd10b9c282745dfd55ef8284d
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>