omx: OMX video decoder changes to co-exist with encoder
In the current design, omx video decoder is sending buffer
ptr for the inbufs and dma_buf_fd for the outbufs to the dce client.
In order to co-exist with encoder the inbufs/outbufs interpretation
is unified between encoder and decoder. this is achieved by
passing the memheader for both inbufs and outbufs for
both encoder and decoders.

Note# This patch needs to be merged along with libdce patch:
http://review.omapzoom.org/#/c/34415/

Change-Id: I646e7bc47f827ab19a3a44a0566b76faa24acfe3
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>