omx: Add TILER2D input buffer support for H264 Video Encoder
This patch adds the support for TILER2D input buffer
configuration for H264 Video Encoder.
Since the omx buffers on i/p port are still non-tiler
there is a memcpy into TILER2D buffers.
This is just an intermediate solution to test TILER2D input
encoding. In the final solution, the memcpy will be removed,
and the TILER handles will be exposed to omx port level.

Change-Id: I0e47af240a729f0bccd33a39c4bb5ff9e42cc374
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>